refactor: Move capture template compiling to template class
